Home_work_16-17(6-7)

+++ VK_application/Models/UsersGetModel.swift
@@ -151,7 +151,7 @@ struct UsersGetLastSeen: Codable {
     /// 6 — приложение для Windows 10;
     /// 7 — полная версия сайта.
     let platform: Int?
-    let time: Int?
+    let time: Double?
 }
 
 struct UsersGetPersonal: Codable {
